+++ test.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+export GR_SATELLITES_SUBMIT_TLM=0
+
 echo 1KUNS-PF
 gr_satellites 1KUNS-PF --wavfile satellite-recordings/1kuns_pf.wav
 echo 3CAT-1